+++ src/pages/api/experiences/[id].ts
@@ -1,4 +1,4 @@
-import { middleware } from 'common/server';
+import { generateId, middleware } from 'common/server';
 import { decamelizeKeys } from 'humps';
 import { JwtPayload } from 'jsonwebtoken';
 import { NextApiRequest, NextApiResponse } from 'next';
@@ -16,6 +16,7 @@ export const ExperienceFormSchema = Yup.object({
   waktu_mulai: Yup.date().nullable().default(null),
   waktu_selesai: Yup.date().nullable().default(null),
   skills: Yup.array(Yup.string().default('')).default([]),
+  lampiran: Yup.array(Yup.string().default('')).default([]),
 });
 
 export default async function handler(
@@ -44,10 +45,11 @@ export default async function handler(
       });
     }
     const user = (await middleware(req, res)) as JwtPayload;
+    const isAdmin = user.type === 'admin';
 
     const nomor_identitas = user.nomor_identitas;
 
-    if (nomor_identitas !== pengalaman.nomorIdentitasMahasiswa) {
+    if (nomor_identitas !== pengalaman.nomorIdentitasMahasiswa && !isAdmin) {
       return res.status(403).json({
         message: 'Anda tidak di-izinkan mengakses fitur ini',
       });
@@ -56,9 +58,23 @@ export default async function handler(
     if (method === 'PUT') {
       const pengalaman = await ExperienceFormSchema.validate(body);
 
-      await prisma.lampiranPengalaman.deleteMany({
-        where: { pengalamanId: id },
-      });
+      if (pengalaman.lampiran.length) {
+        await prisma.$transaction([
+          prisma.lampiranPengalaman.deleteMany({
+            where: { pengalamanId: id },
+          }),
+          prisma.lampiranPengalaman.createMany({
+            data: pengalaman.lampiran.map((file) => {
+              return {
+                fileUrl: file,
+                id: generateId(),
+                jenisFile: 'application/pdf',
+                pengalamanId: id,
+              };
+            }),
+          }),
+        ]);
+      }
 
       const currentExperience = await prisma.pengalaman.update({
         data: {
@@ -69,11 +85,6 @@ export default async function handler(
           skills: pengalaman.skills.join('|'),
           tanggalMulai: pengalaman.waktu_mulai,
           tanggalSelesai: pengalaman.waktu_selesai,
-          LampiranPengalaman: {
-            createMany: {
-              data: [],
-            },
-          },
         },
         where: { id },
         select: PengalamanResouceModel,
